modules moved to their own directory. mostly works (?)
